reused existing util for implementation of linalg.norm
